Pin setuptools<60.0.0
Given that it was "python setup.py testr ..." that was hanging, and before showing any test output,
I guessed the problem might be caused by a setuptools change, and so tried pinning to various
earlier versions.

Without any pin, we were getting setuptools 61.0.0.  This wasn't visible in the Semaphore logs, but
I attached to the Semaphore VM and checked what was installed under the py38 venv.

Then, empirically:

- Pin setuptools<61.0.0: we still get a hang

- Pin setuptools<58.0.0: no hang, test runs through normally

- Pin setuptools<60.0.0: no hang, test runs through normally

So pinning <60.0.0 as it feels better to pin to as recent a version as possible.

Unfortunately I don't see anything, from v59 to v60 at
https://setuptools.pypa.io/en/latest/history.html, that obviously explains this hang, and it's not
ideal to do this pin without deeper understanding.  But it gets our CI green again for now.
Hopefully we'll have a chance to come back and understand better.