Always program OpenStack floating IPs
Felix's programming of NAT mappings was disabled by default in
https://github.com/projectcalico/calico/pull/5861.  This was a security fix for Kubernetes, as in
Kubernetes anyone could create a floating IP for themselves by adding a "floatingips" annotation to
their own pods.

For OpenStack, pod annotations obviously do not apply, and NAT mappings come from the ML2 driver
running as part of the Neutron server, which is already access-protected in an appropriate way.  And
we don't want to have to document that OpenStack users need to add "FloatingIPs = Enabled" to their
felix.cfg file.

Therefore, arrange for OpenStack-configured floating IPs to be programmed regardless of the
FloatingIPs setting in FelixConfiguration.