DHCP agent: Fix check for linux interface being up
We were looking for "state UP" in the output from `ip monitor link`, but that's not the right thing
to look for, and in practice often shows as "state UNKNOWN" when the interface is actually up.

According to https://superuser.com/questions/1463162/unknown-state-of-network-interface, the
correct thing is to look for "UP" in the set of flags between < and >, as for example here:

    3: wlp59s0: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc noqueue state UP mode DORMANT group default qlen 1000

I've adjusted the rx_up regexp to do that.