networking_calico: Do not checksum DHCP packets
Very old versions of isc-dhcp-client broke when UDP
packets had a checksum field not properly filled in as
part of GSO/checksum offload support being introduced
for virtio_net:
https://lwn.net/Articles/373209/

The missing/incorrect checksum in the UDP header was
problematic for isc-dhcp-client (and possibly other DHCP
clients) because they use raw sockets, and were unaware
of how to handle TP_STATUS_CSUMNOTREADY.

OpenStack has historically worked around this by adding an
iptables mangle rule to always rewrite the UDP header for
DHCP packets. But, isc-dhcp-client has also been patched to
account for this since 2014, and likewise has qemu (for
virtio_net) since 2009:
https://git.qemu.org/?p=qemu.git;a=commit;h=1d41b0c

The consequence of patching the mangle rule in iptables is
that the underlying OpenStack mechanisms tend to add the
rules in an order that upsets Felix. When Felix sees this
(often during live-migration, when the VM has no
networking), it can spend a fair amount of time resyncing
the rules... thus leaving a live-migrating instance with
no networking for an appreciable amount of time.

As this should not be needed for any Linux distribution in
the last decade and it can perturb Felix, simply do not
fill in DHCP UDP checksums here with the assumption that
it has been fixed/is being done elsewhere.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>