Use Python package constraints for Ussuri, except for etcd3gw
etcd3gw is an exception because in that case we want to use a later version.

Adding these constraints was specifically prompted by noticing that the latest dnspython (2.3,
released end December) breaks us; but then I realized that we really ought to be using the full set
of constraints that OpenStack curate for each release.